Rename biodata to biometadata, replace Avro mentions
