+++ cts-java/src/test/java/org/ga4gh/cts/api/bioMetadata/BioMetadataTests.java
@@ -1,0 +1,5 @@
+package org.ga4gh.cts.api.bioMetadata;
+
+
+public interface BioMetadataTests {  /* category marker */
+}
