--- cts-java/src/test/java/org/ga4gh/cts/api/biodata/BiodataTests.java
@@ -1,5 +1,0 @@
-package org.ga4gh.cts.api.biodata;
-
-
-public interface BiodataTests {  /* category marker */
-}
